Move Guard and HostList to pages
