+++ assets/js/pages/Guard/index.js
@@ -1,0 +1,3 @@
+import Guard from './Guard';
+
+export default Guard;
