+++ assets/js/pages/HostList/index.js
@@ -1,0 +1,3 @@
+import HostList from './HostList';
+
+export default HostList;
